device-tree: Fix uart2 (ser0) communication
We saw an issue with uart2 hogging the SysB RS485 communication line.
Thus, the HC12 on SysB can't communicate with the controllerbox.
To fix that, revert modifications in #8064565c9bb6 and set up uart2
similar to uart3, which has the equivalent function on SysA.

TODO: Move all modifications to overlay.